fix: satisfy pkg-file-count and backend-size limits for provider package
