+++ tests/functional/providers/codex_failure_reporting_agreement_test.go
@@ -1,26 +1,26 @@
-package provider
+package providers
 
 import (
 	"context"
-	"encoding/json"
 	"strings"
 	"testing"
 
 	"github.com/portpowered/infinite-you/pkg/interfaces"
+	"github.com/portpowered/infinite-you/pkg/workers/provider"
 )
 
 type codexFailureReportingFacts struct {
-	Type            interfaces.WorkFailureType
-	Family          interfaces.WorkFailureFamily
-	Message         string
-	Retryable       bool
-	Terminal        bool
-	ThrottlePause   bool
+	Type          interfaces.WorkFailureType
+	Family        interfaces.WorkFailureFamily
+	Message       string
+	Retryable     bool
+	Terminal      bool
+	ThrottlePause bool
 }
 
-func codexFailureReportingFactsFromResult(result ProviderFailureResult) codexFailureReportingFacts {
-	providerErr := NewProviderErrorFromResult(result, nil)
-	decision := WorkFailureDecisionFromProviderError(providerErr)
+func codexFailureReportingFactsFromResult(result provider.ProviderFailureResult) codexFailureReportingFacts {
+	providerErr := provider.NewProviderErrorFromResult(result, nil)
+	decision := provider.WorkFailureDecisionFromProviderError(providerErr)
 	return codexFailureReportingFacts{
 		Type:          providerErr.Type,
 		Family:        providerErr.Family,
@@ -31,26 +31,11 @@ func codexFailureReportingFactsFromResult(result ProviderFailureResult) codexFai
 	}
 }
 
-func assertCodexFailureReportingFactsEqual(t *testing.T, structuredPath, processExitPath string, got, want codexFailureReportingFacts) {
+func assertCodexFailureReportingFactsEqual(t *testing.T, path string, got, want codexFailureReportingFacts) {
 	t.Helper()
 	if got != want {
-		t.Fatalf("%s facts = %#v, want %#v", structuredPath, got, want)
+		t.Fatalf("%s facts = %#v, want %#v", path, got, want)
 	}
-}
-
-func codexStructuredStreamStdout(message string) []byte {
-	record, err := json.Marshal(map[string]any{
-		"type":  "turn.failed",
-		"error": map[string]string{"message": message},
-	})
-	if err != nil {
-		panic(err)
-	}
-	return append(record, '\n')
-}
-
-func codexProcessExitResult(stderr string, exitCode int) CommandResult {
-	return CommandResult{ExitCode: exitCode, Stderr: []byte(stderr)}
 }
 
 type codexFailureReportingAgreementCase struct {
@@ -139,16 +124,16 @@ func codexFailureReportingAgreementCases() []codexFailureReportingAgreementCase 
 
 func runCodexFailureReportingAgreementCase(t *testing.T, tc codexFailureReportingAgreementCase) {
 	t.Helper()
-	streamResult, ok := CodexStructuredStreamReportingOutcome(codexStructuredStreamStdout(tc.streamMessage))
+	streamResult, ok := provider.CodexStructuredStreamReportingOutcome(codexStructuredStreamStdout(tc.streamMessage))
 	if !ok {
 		t.Fatal("CodexStructuredStreamReportingOutcome() ok = false, want true")
 	}
-	exitResult := CodexProcessExitReportingOutcome(codexProcessExitResult(tc.exitStderr, tc.exitCode))
+	exitResult := provider.CodexProcessExitReportingOutcome(codexProcessExitResult(tc.exitStderr, tc.exitCode))
 
 	streamFacts := codexFailureReportingFactsFromResult(streamResult)
 	exitFacts := codexFailureReportingFactsFromResult(exitResult)
-	assertCodexFailureReportingFactsEqual(t, "structured-stream", "process-exit", streamFacts, tc.want)
-	assertCodexFailureReportingFactsEqual(t, "process-exit", "structured-stream", exitFacts, tc.want)
+	assertCodexFailureReportingFactsEqual(t, "structured-stream", streamFacts, tc.want)
+	assertCodexFailureReportingFactsEqual(t, "process-exit", exitFacts, tc.want)
 	if streamFacts != exitFacts {
 		t.Fatalf("structured-stream facts = %#v, process-exit facts = %#v, want matching outcomes", streamFacts, exitFacts)
 	}
@@ -168,18 +153,18 @@ func TestCodexFailureReportingPaths_AgreeOnListedFailureClasses(t *testing.T) {
 func TestCodexFailureReportingPaths_CanceledAgreesAcrossCompetingSignals(t *testing.T) {
 	competing := []struct {
 		name   string
-		result CommandResult
+		result provider.CommandResult
 	}{
 		{
 			name: "structured_stream_competition",
-			result: CommandResult{
+			result: provider.CommandResult{
 				ExitCode: 1,
 				Stdout:   codexStructuredStreamStdout("unexpected status 429"),
 			},
 		},
 		{
 			name: "process_exit_competition",
-			result: CommandResult{
+			result: provider.CommandResult{
 				ExitCode: 1,
 				Stderr:   []byte("ERROR: unexpected status 401\n"),
 			},
@@ -190,11 +175,11 @@ func TestCodexFailureReportingPaths_CanceledAgreesAcrossCompetingSignals(t *test
 		Type: interfaces.WorkFailureTypeUnknown, Family: interfaces.WorkFailureFamilyTerminal,
 		Message: "Codex execution was canceled.", Terminal: true,
 	}
-	input := CodexFailureResolutionInput{FlushReason: CodexFlushReasonCanceled}
+	input := provider.CodexFailureResolutionInput{FlushReason: provider.CodexFlushReasonCanceled}
 
 	for _, tc := range competing {
 		t.Run(tc.name, func(t *testing.T) {
-			resolved, ok := ResolveCodexProviderFailure(tc.result, input)
+			resolved, ok := provider.ResolveCodexProviderFailure(tc.result, input)
 			if !ok {
 				t.Fatal("ResolveCodexProviderFailure() ok = false, want true")
 			}
@@ -209,24 +194,24 @@ func TestCodexFailureReportingPaths_CanceledAgreesAcrossCompetingSignals(t *test
 func TestCodexFailureReportingPaths_TimeoutAgreesAcrossCompetingSignals(t *testing.T) {
 	competing := []struct {
 		name   string
-		result CommandResult
-		input  CodexFailureResolutionInput
+		result provider.CommandResult
+		input  provider.CodexFailureResolutionInput
 	}{
 		{
 			name: "structured_stream_competition",
-			result: CommandResult{
+			result: provider.CommandResult{
 				ExitCode: 1,
 				Stdout:   codexStructuredStreamStdout("unexpected status 429"),
 			},
-			input: CodexFailureResolutionInput{CommandError: context.DeadlineExceeded},
+			input: provider.CodexFailureResolutionInput{CommandError: context.DeadlineExceeded},
 		},
 		{
 			name: "process_exit_competition",
-			result: CommandResult{
+			result: provider.CommandResult{
 				ExitCode: 124,
 				Stderr:   []byte("ERROR: unexpected status 401\n"),
 			},
-			input: CodexFailureResolutionInput{CommandError: context.DeadlineExceeded},
+			input: provider.CodexFailureResolutionInput{CommandError: context.DeadlineExceeded},
 		},
 	}
 
@@ -237,7 +222,7 @@ func TestCodexFailureReportingPaths_TimeoutAgreesAcrossCompetingSignals(t *testi
 
 	for _, tc := range competing {
 		t.Run(tc.name, func(t *testing.T) {
-			resolved, ok := ResolveCodexProviderFailure(tc.result, tc.input)
+			resolved, ok := provider.ResolveCodexProviderFailure(tc.result, tc.input)
 			if !ok {
 				t.Fatal("ResolveCodexProviderFailure() ok = false, want true")
 			}
