+++ tests/functional/providers/codex_failure_test_constants.go
@@ -1,0 +1,10 @@
+package providers
+
+const (
+	codexAuthFailureMessage       = "Codex authentication failed."
+	codexBadRequestFailureMessage = "Codex rejected the request as invalid."
+	codexServerFailureMessage     = "Codex encountered a temporary server error."
+	codexThrottleFailureMessage   = "Codex is temporarily unavailable due to usage or capacity limits."
+	codexTimeoutFailureMessage    = "Codex request timed out."
+	codexUnknownFailureMessage    = "Codex reported a terminal error."
+)
